FAC_NEWP placed elsewhere
FAC_NEWP only seemed to have something to do with ease of use, so I removed it from the physics (and put it where the player creates planets). This way velocities can just be assigned to a planet directly, instead of through two conversions that would cancel each other out.